refactor: move Paramore.Darker.Policies sources into core (#321)
Tidy First (structural-only): relocate Retryable* attributes and decorators,
Constants, and Polly-based builder extensions from the side package
`Paramore.Darker.Policies` into core under `src/Paramore.Darker/Policies/
{Attributes,Handlers}/`. Side-package csproj left empty pending Step 9
cleanup.
