+++ src/Paramore.Darker/Policies/Attributes/RetryableQueryAttribute.cs
@@ -1,13 +1,14 @@
 using System;
+using Paramore.Darker.Policies.Handlers;
 
-namespace Paramore.Darker.Policies
+namespace Paramore.Darker.Policies.Attributes
 {
     [AttributeUsage(AttributeTargets.Method)]
-    public sealed class RetryableQueryAttributeAsync : QueryHandlerAttributeAsync
+    public sealed class RetryableQueryAttribute : QueryHandlerAttribute
     {
         private readonly string _policyName;
 
-        public RetryableQueryAttributeAsync(int step, string policyName = Constants.RetryPolicyName)
+        public RetryableQueryAttribute(int step, string policyName = Constants.RetryPolicyName)
             : base(step)
         {
             _policyName = policyName;
@@ -20,7 +21,7 @@ namespace Paramore.Darker.Policies
 
         public override Type GetDecoratorType()
         {
-            return typeof(RetryableQueryDecoratorAsync<,>);
+            return typeof(RetryableQueryDecorator<,>);
         }
     }
 }
