+++ src/Paramore.Darker/Policies/Handlers/RetryableQueryDecorator.cs
@@ -5,7 +5,7 @@ using Paramore.Darker.Logging;
 using Polly;
 using Polly.Registry;
 
-namespace Paramore.Darker.Policies
+namespace Paramore.Darker.Policies.Handlers
 {
     public class RetryableQueryDecorator<TQuery, TResult> : IQueryHandlerDecorator<TQuery, TResult>
         where TQuery : IQuery<TResult>
